requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.10
requests==2.32.4
python-dotenv==1.1.1